+++ shoop/addons/admin_module/views/reload.py
@@ -10,11 +10,16 @@ from __future__ import unicode_literals
 import time
 
 from django import forms
+from django.conf import settings
+from django.core.management import call_command
 from django.http.response import HttpResponse, JsonResponse
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import FormView
+from six import StringIO
 
+from shoop.addons.manager import get_enabled_addons
 from shoop.addons.reloader import get_reload_method_classes
+from shoop.apps.settings import reload_apps
 from shoop.utils.excs import Problem
 from shoop.utils.iterables import first
 
@@ -44,6 +49,20 @@ class ReloadMethodForm(forms.Form):
         return first(rm for rm in self.reload_methods if rm.identifier == self.cleaned_data["reload_method"])
 
 
+def finalize_installation_for_enabled_apps():
+    out = StringIO()
+    enabled_addons = get_enabled_addons(settings.SHOOP_ENABLED_ADDONS_FILE)
+    new_apps = [app for app in enabled_addons if app not in settings.INSTALLED_APPS]
+    if new_apps:
+        out.write("Enabling new addons: %s" % new_apps)
+        settings.INSTALLED_APPS += type(settings.INSTALLED_APPS)(new_apps)
+        reload_apps()
+
+    call_command("migrate", "--noinput", stdout=out)
+    call_command("collectstatic", "--noinput", stdout=out)
+    return out.getvalue()
+
+
 class ReloadView(FormView):
     template_name = "shoop/admin/addons/reload.jinja"
     form_class = ReloadMethodForm
@@ -56,4 +75,6 @@ class ReloadView(FormView):
     def get(self, request, *args, **kwargs):
         if request.GET.get("ping"):
             return JsonResponse({"pong": time.time()})
+        elif request.GET.get("finalize"):
+            return JsonResponse({"message": finalize_installation_for_enabled_apps()})
         return super(ReloadView, self).get(request, *args, **kwargs)