+++ shoop/apps/settings.py
@@ -105,3 +105,13 @@ def validate_templates_configuration():
         "The `django_jinja` template backend was not encountered in your TEMPLATES configuration. "
         "See the Shoop or Django-Jinja documentation on more information how to configure things correctly."
     )
+
+
+def reload_apps():
+    import django
+    _KNOWN_SETTINGS.clear()
+    django.apps.apps.app_configs.clear()
+    django.apps.apps.ready = False
+    django.setup()
+    # Rebuild any AppDirectoriesFinder instance.
+    django.contrib.staticfiles.finders.get_finder.cache_clear()