+++ doc/changelog.rst
@@ -76,6 +76,11 @@ General/miscellaneous
 
 - Add campaigns module
 
+Addons
+~~~~~~
+
+- Enable upgrade, migrations and collectstatic from admin
+
 
 Version 3.0.0
 -------------